requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
_openmp_mutex=5.1=1_gnu
_tflow_select=2.3.0=mkl
absl-py=1.1.0=pypi_0
aiohttp=3.8.1=py39h7f8727e_1
aiosignal=1.2.0=pyhd3eb1b0_0
argon2-cffi=21.3.0=pyhd3eb1b0_0
argon2-cffi-bindings=21.2.0=py39h7f8727e_0
astor=0.8.1=py39h06a4308_0
asttokens=2.0.5=pyhd3eb1b0_0
astunparse=1.6.3=py_0
async-timeout=4.0.2=pypi_0
attrs=21.4.0=pyhd3eb1b0_0
backcall=0.2.0=pyhd3eb1b0_0
beautifulsoup4=4.11.1=py39h06a4308_0
blas=1.0=mkl
bleach=4.1.0=pyhd3eb1b0_0
blinker=1.4=py39h06a4308_0
bottleneck=1.3.4=py39hce1f21e_0
brotli=1.0.9=he6710b0_2
brotlipy=0.7.0=py39h27cfd23_1003
bzip2=1.0.8=h7b6447c_0
c-ares=1.18.1=h7f8727e_0
ca-certificates=2022.4.26=h06a4308_0
cachetools=5.2.0=pypi_0
certifi=2022.5.18.1=py39h06a4308_0
cffi=1.15.0=py39hd667e15_1
chardet=4.0.0=pypi_0
charset-normalizer=2.0.4=pyhd3eb1b0_0
click=8.0.4=py39h06a4308_0
commentjson=0.9.0=pypi_0
cryptography=3.4.8=py39hd23ed53_0
cudatoolkit=11.3.1=h2bc3f7f_2
cycler=0.11.0=pyhd3eb1b0_0
cython=0.29.28=py39h295c915_0
dataclasses=0.8=pyh6d0b6a4_7
dbus=1.13.18=hb2f20db_0
debugpy=1.5.1=py39h295c915_0
decorator=5.1.1=pyhd3eb1b0_0
defusedxml=0.7.1=pyhd3eb1b0_0
entrypoints=0.4=py39h06a4308_0
executing=0.8.3=pyhd3eb1b0_0
expat=2.4.4=h295c915_0
fastdtw=0.3.4=pypi_0
ffmpeg=4.2.2=h20bf706_0
flatbuffers=1.12=pypi_0
fontconfig=2.13.1=h6c09931_0
fonttools=4.25.0=pyhd3eb1b0_0
freetype=2.11.0=h70c0345_0
frozenlist=1.3.0=pypi_0
fsspec=2022.5.0=pypi_0
gast=0.4.0=pyhd3eb1b0_0
gensim=4.1.2=py39h295c915_0
giflib=5.2.1=h7b6447c_0
glib=2.69.1=h4ff587b_1
gmp=6.2.1=h295c915_3
gnutls=3.6.15=he1e5248_0
google-auth=2.7.0=pypi_0
google-auth-oauthlib=0.4.6=pypi_0
google-pasta=0.2.0=pyhd3eb1b0_0
grpcio=1.46.3=pypi_0
gst-plugins-base=1.14.0=h8213a91_2
gstreamer=1.14.0=h28cd5cc_2
h5py=3.7.0=pypi_0
hdf5=1.10.6=hb1b8bf9_0
icu=58.2=he6710b0_3
idna=3.3=pyhd3eb1b0_0
importlib-metadata=4.11.4=pypi_0
intel-openmp=2021.4.0=h06a4308_3561
ipykernel=6.9.1=py39h06a4308_0
ipython=8.3.0=py39h06a4308_0
ipython_genutils=0.2.0=pyhd3eb1b0_1
jedi=0.18.1=py39h06a4308_1
jinja2=3.0.3=pyhd3eb1b0_0
joblib=1.1.0=pyhd3eb1b0_0
jpeg=9e=h7f8727e_0
jsonschema=4.4.0=py39h06a4308_0
jupyter_client=7.2.2=py39h06a4308_0
jupyter_contrib_core=0.3.3=py_2
jupyter_contrib_nbextensions=0.5.1=pyhd8ed1ab_2
jupyter_core=4.10.0=py39h06a4308_0
jupyter_highlight_selected_word=0.2.0=py39hf3d152e_1005
jupyter_latex_envs=1.4.6=pyhd8ed1ab_1002
jupyter_nbextensions_configurator=0.4.1=py39hf3d152e_2
jupyterlab_pygments=0.1.2=py_0
keras=2.9.0=pypi_0
keras-preprocessing=1.1.2=pyhd3eb1b0_0
kiwisolver=1.4.2=py39h295c915_0
lame=3.100=h7b6447c_0
lark-parser=0.7.8=pypi_0
lcms2=2.12=h3be6417_0
ld_impl_linux-64=2.38=h1181459_1
libclang=14.0.1=pypi_0
libffi=3.3=he6710b0_2
libgcc-ng=11.2.0=h1234567_1
libgfortran-ng=7.5.0=ha8ba4b0_17
libgfortran4=7.5.0=ha8ba4b0_17
libgomp=11.2.0=h1234567_1
libiconv=1.16=h7f8727e_2
libidn2=2.3.2=h7f8727e_0
libopus=1.3.1=h7b6447c_0
libpng=1.6.37=hbc83047_0
libprotobuf=3.20.1=h4ff587b_0
libsodium=1.0.18=h7b6447c_0
libstdcxx-ng=11.2.0=h1234567_1
libtasn1=4.16.0=h27cfd23_0
libtiff=4.2.0=h2818925_1
libunistring=0.9.10=h27cfd23_0
libuuid=1.0.3=h7f8727e_2
libuv=1.40.0=h7b6447c_0
libvpx=1.7.0=h439df22_0
libwebp=1.2.2=h55f646e_0
libwebp-base=1.2.2=h7f8727e_0
libxcb=1.15=h7f8727e_0
libxml2=2.9.14=h74e7548_0
libxslt=1.1.35=h4e12654_0
lxml=4.8.0=py39h1f438cf_0
lz4-c=1.9.3=h295c915_1
markdown=3.3.7=pypi_0
markupsafe=2.1.1=py39h7f8727e_0
matplotlib=3.5.1=py39h06a4308_1
matplotlib-base=3.5.1=py39ha18d171_1
matplotlib-inline=0.1.2=pyhd3eb1b0_2
mistune=0.8.4=py39h27cfd23_1000
mkl=2021.4.0=h06a4308_640
mkl-service=2.4.0=py39h7f8727e_0
mkl_fft=1.3.1=py39hd3c417c_0
mkl_random=1.2.2=py39h51133e4_0
multidict=6.0.2=pypi_0
munkres=1.1.4=py_0
nbclient=0.5.13=py39h06a4308_0
nbconvert=6.4.4=py39h06a4308_0
nbformat=5.3.0=py39h06a4308_0
ncurses=6.3=h7f8727e_2
nest-asyncio=1.5.5=py39h06a4308_0
nettle=3.7.3=hbbd107a_1
networkit=10.0=pypi_0
networkx=2.8.4=pypi_0
notebook=6.4.11=py39h06a4308_0
numexpr=2.8.1=py39h6abb31d_0
numpy=1.22.3=py39he7a7128_0
numpy-base=1.22.3=py39hf524024_0
oauthlib=3.2.0=pyhd3eb1b0_0
openh264=2.1.1=h4ff587b_0
openssl=1.1.1o=h7f8727e_0
opt_einsum=3.3.0=pyhd3eb1b0_1
packaging=21.3=pyhd3eb1b0_0
pandas=1.4.2=py39h295c915_0
pandocfilters=1.5.0=pyhd3eb1b0_0
parso=0.8.3=pyhd3eb1b0_0
pcre=8.45=h295c915_0
pexpect=4.8.0=pyhd3eb1b0_3
pickleshare=0.7.5=pyhd3eb1b0_1003
pillow=9.0.1=py39h22f2fdc_0
pip=21.2.4=py39h06a4308_0
prometheus_client=0.13.1=pyhd3eb1b0_0
prompt-toolkit=3.0.20=pyhd3eb1b0_0
protobuf=3.19.4=pypi_0
ptyprocess=0.7.0=pyhd3eb1b0_2
pure_eval=0.2.2=pyhd3eb1b0_0
pyasn1=0.4.8=pyhd3eb1b0_0
pyasn1-modules=0.2.8=pypi_0
pycparser=2.21=pyhd3eb1b0_0
pydeprecate=0.3.2=pypi_0
pyg=2.0.4=py39_torch_1.11.0_cu113
pygments=2.11.2=pyhd3eb1b0_0
pyjwt=2.1.0=py39h06a4308_0
pyopenssl=21.0.0=pyhd3eb1b0_1
pyparsing=3.0.4=pyhd3eb1b0_0
pyqt=5.9.2=py39h2531618_6
pyrsistent=0.18.0=py39heee7806_0
pysocks=1.7.1=py39h06a4308_0
python=3.9.12=h12debd9_1
python-dateutil=2.8.2=pyhd3eb1b0_0
python-fastjsonschema=2.15.1=pyhd3eb1b0_0
python-flatbuffers=2.0=pyhd3eb1b0_0
python-louvain=0.15=pyhd3eb1b0_0
python_abi=3.9=2_cp39
pytorch=1.11.0=py3.9_cuda11.3_cudnn8.2.0_0
pytorch-cluster=1.6.0=py39_torch_1.11.0_cu113
pytorch-lightning=1.6.4=pypi_0
pytorch-mutex=1.0=cuda
pytorch-scatter=2.0.9=py39_torch_1.11.0_cu113
pytorch-sparse=0.6.13=py39_torch_1.11.0_cu113
pytorch-spline-conv=1.2.1=py39_torch_1.11.0_cu113
pytz=2021.3=pyhd3eb1b0_0
pyyaml=6.0=py39h7f8727e_1
pyzmq=22.3.0=py39h295c915_2
qt=5.9.7=h5867ecd_1
readline=8.1.2=h7f8727e_1
requests=2.27.1=pyhd3eb1b0_0
requests-oauthlib=1.3.1=pypi_0
rsa=4.8=pypi_0
scikit-learn=1.0.2=py39h51133e4_1
scipy=1.8.1=pypi_0
send2trash=1.8.0=pyhd3eb1b0_1
setuptools=61.2.0=py39h06a4308_0
sip=4.19.13=py39h295c915_0
six=1.16.0=pyhd3eb1b0_1
smart_open=5.2.1=py39h06a4308_0
snap-stanford=6.0.0=pypi_0
soupsieve=2.3.1=pyhd3eb1b0_0
sqlite=3.38.3=hc218d9a_0
stack_data=0.2.0=pyhd3eb1b0_0
stellargraph=1.2.1=py_0
tensorboard=2.9.0=pypi_0
tensorboard-data-server=0.6.1=pypi_0
tensorboard-plugin-wit=1.8.1=pypi_0
tensorflow=2.9.1=pypi_0
tensorflow-base=2.4.1=mkl_py39h43e0292_0
tensorflow-estimator=2.9.0=pypi_0
tensorflow-io-gcs-filesystem=0.26.0=pypi_0
termcolor=1.1.0=pypi_0
terminado=0.13.1=py39h06a4308_0
testpath=0.5.0=pyhd3eb1b0_0
threadpoolctl=2.2.0=pyh0d69192_0
tk=8.6.12=h1ccaba5_0
torchmetrics=0.9.0=pypi_0
torchvision=0.12.0=py39_cu113
tornado=6.1=py39h27cfd23_0
tqdm=4.64.0=py39h06a4308_0
traitlets=5.1.1=pyhd3eb1b0_0
typing-extensions=4.1.1=hd3eb1b0_0
typing_extensions=4.1.1=pyh06a4308_0
tzdata=2022a=hda174b7_0
urllib3=1.26.9=py39h06a4308_0
wcwidth=0.2.5=pyhd3eb1b0_0
webencodings=0.5.1=py39h06a4308_1
werkzeug=2.1.2=pypi_0
wheel=0.37.1=pyhd3eb1b0_0
wrapt=1.14.1=pypi_0
x264=1!157.20191217=h7b6447c_0
xz=5.2.5=h7f8727e_1
yacs=0.1.6=pyhd3eb1b0_1
yaml=0.2.5=h7b6447c_0
yarl=1.7.2=pypi_0
zeromq=4.3.4=h2531618_0
zipp=3.8.0=py39h06a4308_0
zlib=1.2.12=h7f8727e_2
zstd=1.5.2=ha4553b6_0
